added action and reducer for adding a tone
