+++ src/store/reducers/tones.js
@@ -1,13 +1,13 @@
 // ADD AND REMOVE TONE
-let initialState = {};
+let initialState = [];
 
 export default (state = initialState, action) => {
   
   let { type, payload } = action;
 
   switch (type) {
-    case 'ADD_TONE': 
-      return Object.assign({}, state, payload);
+    case 'ADD_TONE':
+      return [...state, payload];
     
     default:
       return state;
